+++ src/utils/MatrixUtils.cpp
@@ -10,11 +10,11 @@ vector<double> MatrixUtils::multMatVec(
 ) {
     assert(mat[0].size() == vec.size());
 
-    int matRows = mat.size();
+    size_t matRows = mat.size();
     vector<double> product(matRows, 0.0);
 
     #pragma omp parallel for
-    for (int i = 0; i < matRows; i++) {
+    for (size_t i = 0; i < matRows; i++) {
         product[i] = dot(mat[i], vec);
     }
 
@@ -27,11 +27,11 @@ double MatrixUtils::dot(
 ) {
     assert(vec1.size() == vec2.size());
 
-    int size = vec1.size();
+    size_t size = vec1.size();
     double product = 0.0;
 
     #pragma omp parallel for reduction(+:product)
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         product += vec1[i] * vec2[i];
     }
 
@@ -44,10 +44,10 @@ void MatrixUtils::addVecInplace(
 ) {
     assert(vec1.size() == vec2.size());
 
-    int size = vec1.size();
+    size_t size = vec1.size();
 
     #pragma omp parallel for
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         vec1[i] += vec2[i];
     }
 }
@@ -56,20 +56,20 @@ vector<vector<double> > MatrixUtils::multMatTMat(
     const vector<vector<double> > &mat1,
     const vector<vector<double> > &mat2
 ) {
-    int mat1TRows = mat1[0].size();
-    int mat1TCols = mat1.size();
-    int mat2Rows = mat2.size();
-    int mat2Cols = mat2[0].size();
+    size_t mat1TRows = mat1[0].size();
+    size_t mat1TCols = mat1.size();
+    size_t mat2Rows = mat2.size();
+    size_t mat2Cols = mat2[0].size();
 
     assert(mat1TCols == mat2Rows);
 
     vector<vector<double> > product(mat1TRows, vector<double>(mat2Cols, 0.0));
 
     #pragma omp parallel for collapse(2)
-    for (int i = 0; i < mat1TRows; i++) {
-        for (int j = 0; j < mat2Cols; j++) {
+    for (size_t i = 0; i < mat1TRows; i++) {
+        for (size_t j = 0; j < mat2Cols; j++) {
             double value = 0.0;
-            for (int k = 0; k < mat2Rows; k++) {
+            for (size_t k = 0; k < mat2Rows; k++) {
                 value += mat1[k][i] * mat2[k][j];
             }
             product[i][j] = value;
@@ -83,20 +83,20 @@ vector<vector<double> > MatrixUtils::multMatMatT(
     const vector<vector<double> > &mat1,
     const vector<vector<double> > &mat2
 ) {
-    int mat1Rows = mat1.size();
-    int mat1Cols = mat1[0].size();
-    int mat2TRows = mat2[0].size();
-    int mat2TCols = mat2.size();
+    size_t mat1Rows = mat1.size();
+    size_t mat1Cols = mat1[0].size();
+    size_t mat2TRows = mat2[0].size();
+    size_t mat2TCols = mat2.size();
 
     assert(mat1Cols == mat2TRows);
 
     vector<vector<double> > product(mat1Rows, vector<double>(mat2TCols, 0.0));
 
     #pragma omp parallel for collapse(2)
-    for (int i = 0; i < mat1Rows; i++) {
-        for (int j = 0; j < mat2TCols; j++) {
+    for (size_t i = 0; i < mat1Rows; i++) {
+        for (size_t j = 0; j < mat2TCols; j++) {
             double value = 0.0;
-            for (int k = 0; k < mat2TRows; k++) {
+            for (size_t k = 0; k < mat2TRows; k++) {
                 value += mat1[i][k] * mat2[j][k];
             }
             product[i][j] = value;
@@ -110,10 +110,10 @@ void MatrixUtils::scaleVecInplace(
     vector<double> &vec,
     double scaleFactor
 ) {
-    int size = vec.size();
+    size_t size = vec.size();
 
     #pragma omp parallel for
-    for (int i = 0; i < size; i++) {
+    for (size_t i = 0; i < size; i++) {
         vec[i] *= scaleFactor;
     }
 }
@@ -121,8 +121,8 @@ void MatrixUtils::scaleVecInplace(
 vector<double> MatrixUtils::colSums(
     const vector<vector<double> > &mat
 ) {
-    int numCols = mat[0].size();
-    int numRows = mat.size(); 
+    size_t numCols = mat[0].size();
+    size_t numRows = mat.size(); 
     vector<double> colSumsVec(numCols, 0.0);
 
     #pragma omp parallel 
@@ -130,15 +130,15 @@ vector<double> MatrixUtils::colSums(
         vector<double> threadColSums(numCols, 0.0);
 
         #pragma omp for
-        for (int i = 0; i < numRows; i++) {
-            for (int j = 0; j < numCols; j++) {
+        for (size_t i = 0; i < numRows; i++) {
+            for (size_t j = 0; j < numCols; j++) {
                 threadColSums[j] += mat[i][j];
             }
         }
         
         #pragma omp critical 
         {
-            for (int j = 0; j < numCols; j++) {
+            for (size_t j = 0; j < numCols; j++) {
                 colSumsVec[j] += threadColSums[j];
             }
         }
@@ -151,19 +151,19 @@ vector<vector<double> > MatrixUtils::multMatMat(
     const vector<vector<double> > &mat1,
     const vector<vector<double> > &mat2
 ) {
-    int mat1Cols = mat1[0].size();
-    int mat1Rows = mat1.size();
-    int mat2Cols = mat2[0].size();
-    int mat2Rows = mat2.size();
+    size_t mat1Cols = mat1[0].size();
+    size_t mat1Rows = mat1.size();
+    size_t mat2Cols = mat2[0].size();
+    size_t mat2Rows = mat2.size();
     assert(mat1Cols == mat2Rows);
 
     vector<vector<double> > product(mat1Rows, vector<double>(mat2Cols, 0.0));
 
     #pragma omp parallel for collapse(2)
-    for (int i = 0; i < mat1Rows; i++) {
-        for (int j = 0; j < mat2Cols; j++) {
+    for (size_t i = 0; i < mat1Rows; i++) {
+        for (size_t j = 0; j < mat2Cols; j++) {
             double value = 0.0;
-            for (int k = 0; k < mat2Rows; k++) {
+            for (size_t k = 0; k < mat2Rows; k++) {
                 value += mat1[i][k] * mat2[k][j];
             }
             product[i][j] = value;
@@ -180,12 +180,12 @@ void MatrixUtils::hardamardInplace(
     assert(mat1[0].size() == mat2[0].size());
     assert(mat1.size() == mat2.size());
 
-    int numRows = mat1.size();
-    int numCols = mat1[0].size();
+    size_t numRows = mat1.size();
+    size_t numCols = mat1[0].size();
 
     #pragma omp parallel for collapse(2)
-    for (int i = 0; i < numRows; i++) {
-        for (int j = 0; j < numCols; j++) {
+    for (size_t i = 0; i < numRows; i++) {
+        for (size_t j = 0; j < numCols; j++) {
             mat1[i][j] *= mat2[i][j];
         }
     }
@@ -195,12 +195,12 @@ void MatrixUtils::scaleMatInplace(
     vector<vector<double> > &mat,
     double scaleFactor
 ) {
-    int numRows = mat.size();
-    int numCols = mat[0].size();
+    size_t numRows = mat.size();
+    size_t numCols = mat[0].size();
 
     #pragma omp parallel for collapse(2)
-    for (int i = 0; i < numRows; i++) {
-        for (int j = 0; j < numCols; j++) {
+    for (size_t i = 0; i < numRows; i++) {
+        for (size_t j = 0; j < numCols; j++) {
             mat[i][j] *= scaleFactor;
         }
     }
@@ -213,13 +213,13 @@ void MatrixUtils::addMatInplace(
     assert(mat1[0].size() == mat2[0].size());
     assert(mat1.size() == mat2.size());
 
-    int numRows = mat1.size();
-    int numCols = mat1[0].size();
+    size_t numRows = mat1.size();
+    size_t numCols = mat1[0].size();
 
 
     #pragma omp parallel for collapse(2)
-    for (int i = 0; i < numRows; i++) {
-        for (int j = 0; j < numCols; j++) {
+    for (size_t i = 0; i < numRows; i++) {
+        for (size_t j = 0; j < numCols; j++) {
             mat1[i][j] += mat2[i][j];
         }
     }
